# Conflicts:
#	src/types/database.d.ts

+++ src/routes/script/updateScript.ts
@@ -20,15 +20,20 @@ export default router.post(
       name,
       content,
     });
-    const assetsData = await u.db("o_assets").whereIn("id", assets).select();
-    await u.db("o_scriptAssets").where({ scriptId: id }).delete();
-    const insertData = assetsData.map((item) => {
-      return {
-        scriptId: id,
-        assetId: item.id,
-      };
-    });
-    await u.db("o_scriptAssets").insert(insertData);
+    if (assets.length) {
+      const assetsData = await u.db("o_assets").whereIn("id", assets).select();
+      await u.db("o_scriptAssets").where({ scriptId: id }).delete();
+      if (assetsData.length) {
+        const insertData = assetsData.map((item) => {
+          return {
+            scriptId: id,
+            assetId: item.id,
+          };
+        });
+        await u.db("o_scriptAssets").insert(insertData);
+      }
+    }
+
     res.status(200).send(success({ message: "编辑剧本成功" }));
   },
 );
